Implemented tracer middle end and created HTTP metrics
Before we begin implementing the core logic of our app, we want some
tooling to help us monitor and make sense of how our app is doing.
To that end we added a tracing middleware that will help us extract info
from all served requests.
We also implemented a prometheus metric for keeping track of response
codes for our server - this will be useful to showcase prometheus-style
metrics.

Worth noticing: if we instrument the application with a distributed
tracing SDK (e.g., honeycomb's beeline) we get both of the things
implemented here [and a lot more useful data] almost for free.
This is just to get an idea of how to monitor our app.

Signed-off-by: fundoplicatedFundus <jorge.alarc123@gmail.com>